+++ packages/axios-extension/src/abap/abap-service-provider.ts
@@ -7,6 +7,7 @@ import { ATO_CATALOG_URL_PATH, parseAtoResponse, TenantType } from './ato';
 import { Ui5AbapRepositoryService } from './ui5-abap-repository-service';
 import { AppIndexService } from './app-index-service';
 import { ODataVersion } from '../base/odata-service';
+import { LayeredRepositoryService } from './lrep-service';
 
 export interface AbapServiceProviderExtension {
     s4Cloud: boolean | undefined;
@@ -136,5 +137,20 @@ export class AbapServiceProvider extends ServiceProvider implements AbapServiceP
             );
         }
         return this.services[AppIndexService.PATH] as AppIndexService;
+    }
+
+    /**
+     * Create or get an existing instance of design time adaptation service.
+     *
+     * @returns an instance of the design time adaptation service.
+     */
+    public layeredRepository(): LayeredRepositoryService {
+        if (!this.services[LayeredRepositoryService.PATH]) {
+            this.services[LayeredRepositoryService.PATH] = this.createService<LayeredRepositoryService>(
+                LayeredRepositoryService.PATH,
+                LayeredRepositoryService
+            );
+        }
+        return this.services[LayeredRepositoryService.PATH] as LayeredRepositoryService;
     }
 }