+++ packages/axios-extension/test/abap/lrep-service.test.ts
@@ -1,0 +1,133 @@
+import nock from 'nock';
+import fs from 'fs';
+import type { Message } from '../../src/abap/lrep-service';
+import { LayeredRepositoryService, createForAbap, AdaptationConfig } from '../../src';
+
+nock.disableNetConnect();
+
+describe('LayeredRepositoryService', () => {
+    const server = 'http://sap.example';
+    const service = createForAbap({ baseURL: server }).layeredRepository();
+    const config: AdaptationConfig = {
+        namespace: 'apps/my.base.app/appVariants/customer.variant/',
+        package: 'MY_PACKAGE',
+        transport: 'MYTRANSPORT'
+    };
+
+    beforeEach(() => {
+        jest.resetModules();
+    });
+
+    afterEach(() => {
+        jest.clearAllMocks();
+    });
+
+    describe('deploy', () => {
+        const archivePath = './dist/my-app.zip';
+        const testData = Buffer.from('TestData');
+
+        beforeAll(() => {
+            jest.spyOn(fs, 'readFileSync').mockReturnValue(testData);
+        });
+
+        test('deploy new adapation project', async () => {
+            nock(server)
+                .get(
+                    `${LayeredRepositoryService.PATH}/dta_folder/?name=${encodeURIComponent(
+                        config.namespace as string
+                    )}&layer=CUSTOMER_BASE`
+                )
+                .reply(404);
+
+            nock(server)
+                .post(
+                    `${LayeredRepositoryService.PATH}/dta_folder/?name=${encodeURIComponent(
+                        config.namespace as string
+                    )}&layer=CUSTOMER_BASE&package=${config.package}&changeList=${config.transport}`
+                )
+                .reply(200);
+
+            const response = await service.deploy(archivePath, config);
+            expect(response.status).toBe(200);
+        });
+
+        test('update an existing adapation project', async () => {
+            nock(server)
+                .get(
+                    `${LayeredRepositoryService.PATH}/dta_folder/?name=${encodeURIComponent(
+                        config.namespace as string
+                    )}&layer=CUSTOMER_BASE`
+                )
+                .reply(200, undefined, {
+                    'x-csrf-token': 'token'
+                });
+
+            nock(server)
+                .put(
+                    `${LayeredRepositoryService.PATH}/dta_folder/?name=${encodeURIComponent(
+                        config.namespace as string
+                    )}&layer=CUSTOMER_BASE&package=${config.package}&changeList=${config.transport}`
+                )
+                .reply(200, {
+                    result: {
+                        id: '~id1',
+                        severity: 'Success',
+                        text: '~text'
+                    } as Message,
+                    messages: [
+                        {
+                            id: '~id2',
+                            severity: 'Warning',
+                            text: '~text'
+                        } as Message
+                    ]
+                });
+
+            const response = await service.deploy(archivePath, config);
+            expect(response.status).toBe(200);
+        });
+    });
+
+    describe('undeploy', () => {
+        test('undeploy existing adapation project', async () => {
+            nock(server)
+                .get(
+                    `${LayeredRepositoryService.PATH}/dta_folder/?name=${encodeURIComponent(
+                        config.namespace as string
+                    )}&layer=CUSTOMER_BASE`
+                )
+                .reply(200, undefined, {
+                    'x-csrf-token': 'token'
+                });
+            nock(server)
+                .delete(
+                    `${LayeredRepositoryService.PATH}/dta_folder/?name=${encodeURIComponent(
+                        config.namespace as string
+                    )}&layer=CUSTOMER_BASE&changeList=${config.transport}`
+                )
+                .reply(200, 'Response that throws an error when given to JSON.parse');
+
+            const response = await service.undeploy(config);
+            expect(response.status).toBe(200);
+        });
+
+        test('try undeploying a not existing adapation project', async () => {
+            nock(server)
+                .get(
+                    `${LayeredRepositoryService.PATH}/dta_folder/?name=${encodeURIComponent(
+                        config.namespace as string
+                    )}&layer=CUSTOMER_BASE`
+                )
+                .reply(404, undefined, {
+                    'x-csrf-token': 'token'
+                });
+
+            try {
+                await service.undeploy(config);
+                fail('The function should have thrown an error.');
+            } catch (error) {
+                expect(error).toBeDefined();
+            }
+        });
+    });
+});