+++ examples/odata-cli/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @sap-ux/odata-cli
 
+## 0.8.3
+
+### Patch Changes
+
+-   Updated dependencies [815bf59]
+    -   @sap-ux/btp-utils@0.9.1
+    -   @sap-ux/axios-extension@0.5.1
+
+## 0.8.2
+
+### Patch Changes
+
+-   Updated dependencies [439b9d0]
+    -   @sap-ux/axios-extension@0.5.0
+
 ## 0.8.1
 
 ### Patch Changes