+++ packages/axios-extension/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @sap-ux/axios-extension
 
+## 0.5.1
+
+### Patch Changes
+
+-   Updated dependencies [815bf59]
+    -   @sap-ux/btp-utils@0.9.1
+
+## 0.5.0
+
+### Minor Changes
+
+-   439b9d0: Added abstraction for LREP service
+
 ## 0.4.0
 
 ### Minor Changes